+++ src/main/java/com/mincho/herb/domain/bookmark/application/HerbBookmarkService.java
@@ -6,7 +6,7 @@ import com.mincho.herb.domain.bookmark.dto.HerbBookmarkResponseDTO;
 public interface HerbBookmarkService {
     void removeHerbBookmark(Long favoriteHerbId);
     void addHerbBookmark(String url, Long herbId);
-    Integer getBookmarkCount(Long herbId);
+    Long getBookmarkCount(Long herbId);
     Boolean isBookmarked(Long herbId);
     HerbBookmarkResponseDTO getBookmarks(int page, int size);
